Add two colors which match the Figma components Move DisplayMetrics from MainActivity to Fragment Add ViewUtil class for functions related to the View/UI Transfer functions from the Fragment to the ViewUtil class
