+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,6 @@
     <color name="colorPrimaryDark">#00574B</color>
     <color name="colorAccent">#D81B60</color>
     <color name="colorInactive">#ACAAAA</color>
+    <color name="colorAdd">#21242C</color>
+    <color name="colorSuggest">#E4E4E4</color>
 </resources>
